+++ src/main/java/admin/adminsiteserver/member/application/MemberService.java
@@ -7,9 +7,9 @@ import admin.adminsiteserver.member.domain.MemberRepository;
 import admin.adminsiteserver.member.exception.MemberAlreadyExistException;
 import admin.adminsiteserver.member.exception.StudentNumberAlreadyExistException;
 import admin.adminsiteserver.member.exception.MemberNotFoundException;
-import admin.adminsiteserver.member.dto.request.SignUpRequest;
-import admin.adminsiteserver.member.dto.request.UpdateMemberRequest;
-import admin.adminsiteserver.member.dto.request.UpdatePasswordRequest;
+import admin.adminsiteserver.member.ui.request.SignUpRequest;
+import admin.adminsiteserver.member.ui.request.UpdateMemberRequest;
+import admin.adminsiteserver.member.ui.request.UpdatePasswordRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
