+++ src/main/java/admin/adminsiteserver/member/ui/MemberController.java
@@ -4,12 +4,12 @@ import admin.adminsiteserver.authentication.util.AuthenticationPrincipal;
 import admin.adminsiteserver.common.dto.LoginMember;
 import admin.adminsiteserver.member.application.MemberQueryService;
 import admin.adminsiteserver.member.application.MemberService;
-import admin.adminsiteserver.member.dto.response.MemberResponse;
-import admin.adminsiteserver.member.dto.request.SignUpRequest;
-import admin.adminsiteserver.member.dto.request.UpdateImageRequest;
-import admin.adminsiteserver.member.dto.request.UpdateMemberRequest;
-import admin.adminsiteserver.member.dto.request.UpdatePasswordRequest;
-import admin.adminsiteserver.member.dto.response.MembersResponse;
+import admin.adminsiteserver.member.ui.response.MemberResponse;
+import admin.adminsiteserver.member.ui.request.SignUpRequest;
+import admin.adminsiteserver.member.ui.request.UpdateImageRequest;
+import admin.adminsiteserver.member.ui.request.UpdateMemberRequest;
+import admin.adminsiteserver.member.ui.request.UpdatePasswordRequest;
+import admin.adminsiteserver.member.ui.response.MembersResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
